Move prometheus api middleware to package
To avoid it being included in the client.
